refactor: split between style and page markup
