+++ web/src/pages/Login/index.tsx
@@ -1,18 +1,10 @@
 import { useState } from 'react';
-import styled from 'styled-components';
-import RegisterForm from '../components/forms/RegisterForm';
-import LoginForm from '../components/forms/LoginForm';
-import ForgotPasswordForm from '../components/forms/ForgotPassword';
-import NavButtonCTA from '../components/buttons/NavButtonCTA';
-import { Container } from '../components/styled';
-
-const TopSection = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  grid-gap: 1rem;
-  padding: 1rem;
-`;
+import RegisterForm from '../../components/forms/RegisterForm';
+import LoginForm from '../../components/forms/LoginForm';
+import ForgotPasswordForm from '../../components/forms/ForgotPassword';
+import NavButtonCTA from '../../components/buttons/NavButtonCTA';
+import { Container } from '../../components/styled';
+import TopSection from './styled';
 
 interface Props {
   setErrorMessage: (message: string) => void;
