+++ web/src/pages/Login/styled.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const TopSection = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  grid-gap: 1rem;
+  padding: 1rem;
+`;
+
+export default TopSection;
